Everything has been moved into the handler package
